Shift slow test related pytest hooks in lower level conftest.py (#578)
* Move slow test related methods to lower level conftest.py

- They conflicted with astropy's test plugins which
  were imported as *.

* Use session wide temp director to store plots.

* Use remote_data marker on pytest_unconfigure.

* Shift unpacking of integration tests config in pytest_configure.

* Cosmetic change - remove unused imports from conftest.py

* Add a check in pytest_unconfigure for integration tests.
